+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Updated/Enhanced Rules and Bug Fixes
  - #357: **PrivateFieldCouldBeFinal** rule: Fix the fasle positive that occurs when a field is marked with the Lazy annotation. (Jenn Strater)
  - #372: **UnnecessaryGetter** rule: Also handle the "Spy" Spock method. (Boris Petrov)
  - #383: **DuplicateImport** rule: Fix handling of imports with non-English characters.
+ - #384: **JUnitLostTest** rule: ignore abstract test methods.
+
  
 Version 1.2.1 (Aug 2018)
 --------------------------------------